Rename Webpack config to official recommendation
